feat: 添加订单和支付状态更新功能
此次提交引入了订单和支付状态更新的功能，包括状态机流转、事件触发和相应的RPC接口。主要修改包括：
1. 在`order.proto`和`payment.proto`中新增状态更新相关的枚举和消息定义。
2. 新增`UpdateOrderStatus`和`UpdatePaymentStatus`的RPC接口及其实现。
3. 更新了订单和支付的状态枚举，支持更多状态流转。
4. 新增了相关的测试文件和业务逻辑处理。

这些改动为订单和支付的状态管理提供了更灵活和可扩展的支持。

+++ app/payment/biz/model/payment.go
@@ -15,7 +15,7 @@ const (
 type Payment struct {
 	gorm.Model
 	OrderId int64 `gorm:"type:int;not null"`
-	Amount float64 `gorm:"type:decimal(10,2);not null"`      // 支付金额
+	Amount float32 `gorm:"type:decimal(10,2);not null"`      // 支付金额
 	Status PaymentStatus `gorm:"type:tinyint;not null;default:1"` 
 	PayChannel string `gorm:"type:varchar(20);not null"` // 支付状态
 	PayURL string `gorm:"type:text;not null"`
